pyproject.toml
[project]
name = "ai-newsletter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "huggingface-hub>=1.8.0",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "uvicorn>=0.42.0",
]
